Store sums as int

+++ internal/core/ports/service.go
@@ -15,7 +15,7 @@ type OrderService interface {
 	AddOrder(ctx context.Context, user *domain.User, orderNumber string) error
 	GetAllOrders(ctx context.Context, user *domain.User) ([]domain.Order, error)
 	GetUserBalance(ctx context.Context, user *domain.User) (domain.UserBalance, error)
-	Withdraw(ctx context.Context, orderNumber string, sum float64, user *domain.User) error
+	Withdraw(ctx context.Context, orderNumber string, sum int, user *domain.User) error
 	GetAllWithdrawals(ctx context.Context, user *domain.User) ([]domain.Withdrawn, error)
 }
